+++ QuanLySieuThi/src/main/java/com/example/quanlysieuthi/service/impl/NhanVienServiceImpl.java
@@ -1,7 +1,11 @@
 package com.example.quanlysieuthi.service.impl;
 
+import com.example.quanlysieuthi.data.entity.HoaDon;
 import com.example.quanlysieuthi.data.entity.NhanVien;
+import com.example.quanlysieuthi.data.entity.PhieuNhap;
+import com.example.quanlysieuthi.data.repository.HoaDonRepository;
 import com.example.quanlysieuthi.data.repository.NhanVienRepository;
+import com.example.quanlysieuthi.data.repository.PhieuNhapRepository;
 import com.example.quanlysieuthi.dto.request.NhanVienRequest;
 import com.example.quanlysieuthi.exceptions.ResourceNotAcceptException;
 import com.example.quanlysieuthi.exceptions.ResourceNotFoundException;
@@ -19,10 +23,14 @@ import java.util.List;
 public class NhanVienServiceImpl implements NhanVienService {
 
     public final NhanVienRepository nhanVienRepository;
+    public final PhieuNhapRepository phieuNhapRepository;
+    public final HoaDonRepository hoaDonRepository;
     public final ModelMapper modelMapper;
 
-    public NhanVienServiceImpl(NhanVienRepository nhanVienRepository, ModelMapper modelMapper) {
+    public NhanVienServiceImpl(NhanVienRepository nhanVienRepository, PhieuNhapRepository phieuNhapRepository, HoaDonRepository hoaDonRepository, ModelMapper modelMapper) {
         this.nhanVienRepository = nhanVienRepository;
+        this.phieuNhapRepository = phieuNhapRepository;
+        this.hoaDonRepository = hoaDonRepository;
         this.modelMapper = modelMapper;
     }
 
@@ -45,6 +53,9 @@ public class NhanVienServiceImpl implements NhanVienService {
     public void creatNhanVien(NhanVienRequest dto) {
         List<NhanVien> nhanVienList = this.nhanVienRepository.findAll();
         LinkedList<NhanVien> linkedList = convertToLinkedList(nhanVienList);
+        if (dto.getCccd().length() != 9 || dto.getCccd().length() != 13){
+            throw new ResourceNotAcceptException("Số căn cước công dân không hợp lệ!!");
+        }
         if (linkedList.isEmpty()){
             NhanVien nhanVien = modelMapper.map(dto, NhanVien.class);
             linkedList.addFirst(nhanVien);
@@ -53,7 +64,7 @@ public class NhanVienServiceImpl implements NhanVienService {
         else {
             for (NhanVien nhanVien : linkedList){
                 if (dto.getMa().equals(nhanVien.getMa())){
-                    throw new ResourceNotAcceptException("Nhan Vien already exist");
+                    throw new ResourceNotAcceptException("Mã nhân viên đã tồn tại!!");
                 }
             }
             NhanVien nhanVien = modelMapper.map(dto, NhanVien.class);
@@ -71,14 +82,39 @@ public class NhanVienServiceImpl implements NhanVienService {
     }
 
     @Override
-    public LinkedList<NhanVien> deleteNhanVien(String ma) {
-        NhanVien nhanVien =nhanVienRepository.findById(ma).get();
+    public void deleteNhanVien(String ma) {
+        NhanVien nhanVien = nhanVienRepository.findById(ma).orElse(null);
+        if (nhanVien == null) {
+            throw new ResourceNotFoundException("Không tìm thấy nhân viên có mã " + ma);
+        }
+
+        List<PhieuNhap> phieuNhapList = this.phieuNhapRepository.findAll();
+        boolean hasPhieuNhap = false;
+        for (PhieuNhap phieuNhap : phieuNhapList) {
+            if (ma.equals(phieuNhap.getSanPham().getMa())) {
+                hasPhieuNhap = true;
+                break;
+            }
+        }
+
+        List<HoaDon> hoaDonList = this.hoaDonRepository.findAll();
+        boolean hasHoaDon = false;
+        for (HoaDon hoaDon : hoaDonList) {
+            if (ma.equals(hoaDon.getSanPham().getMa())) {
+                hasHoaDon = true;
+                break;
+            }
+        }
+
+        if (hasPhieuNhap) {
+            throw new ResourceNotAcceptException("Nhân viên có trong phiếu nhập. Không thể xóa!!!");
+        }
+
+        if (hasHoaDon) {
+            throw new ResourceNotAcceptException("Nhân viên có trong hóa đơn. Không thể xóa!!!");
+        }
+
         nhanVienRepository.delete(nhanVien);
-
-        List<NhanVien> nhanVienList = this.nhanVienRepository.findAll();
-        LinkedList<NhanVien> linkedList = convertToLinkedList(nhanVienList);
-
-        return linkedList;
     }
 
     @Override
